Time: 0 ms (100%), Space: 44.1 MB (24.26%) - LeetHub

+++ 1677-matrix-diagonal-sum/1677-matrix-diagonal-sum.java
@@ -3,9 +3,9 @@ class Solution {
         byte side = (byte)mat.length;
         short sum = 0;
         for(int i = 0; i<side; i++)
-            sum+=(mat[i][i] + mat[i][side-i-1]); 
+            sum+=(short)(mat[i][i] + mat[i][side-i-1]); 
         if(side %2 != 0)
-            sum-= mat[(side/2)][(side/2)];
+            sum-=(short)mat[(side/2)][(side/2)];
         return sum;
     }
 }
